rename devicearray to array
