+++ ivy/functional/frontends/jax/__init__.py
@@ -4,8 +4,8 @@ import sys
 # local
 from ivy.functional.frontends import set_frontend_to_specific_version
 from . import config
-from . import devicearray
-from .devicearray import Array
+from . import array
+from .array import Array
 from . import general_functions
 from .general_functions import *
 from . import lax
